Cast computed_avspeed to float

+++ cycling_data/models/prediction.py
@@ -126,7 +126,7 @@ def prepare_model_dataset(rides,dbsession,predict_columns,extra_fields=[]):
     dataset['equipment']=dataset['equipment_id'].map({equipment.id:equipment.name for equipment in equipments})
 
     rolling_time=pd.to_timedelta(dataset.rolling_time,errors='coerce')
-    computed_avspeed=dataset.distance/rolling_time.dt.total_seconds()*3600
+    computed_avspeed=(dataset.distance/rolling_time.dt.total_seconds()*3600).astype(float)
     pd.options.mode.use_inf_as_na = True
 
     dataset.avspeed.fillna(computed_avspeed,inplace=True)
